Apa ngapain hayoo nyari cewe ?
Commit gas ae bree

+++ lib/messages/converter.ts
@@ -21,22 +21,22 @@ export class Converter {
 		globalThis.CMD.on("converter|sticker", ["sticker", "s", "stiker", "stickergif", "stikergif", "sgif"], async (res: WAConnection, data: Commands) => {
 			const { args, media, from, mess, sendOwner, isQuotedDokumen, Command, FileSha, sender } = data;
 			if (!media ||  isQuotedDokumen) return await this.Ra.reply(from,  IndStickerReply(Command), mess)
-			if (CheckSticker(from, FileSha, Stick)) {
-				const Hasil: { sender: string, id: number, filesha: string, mess: proto.WebMessageInfo}[] = Stick.get(from)
+			if (CheckSticker(from, FileSha || "", Stick)) {
+				const Hasil: { sender: string, id: number, filesha: string, mess: proto.WebMessageInfo}[] | undefined = Stick.get(from)
 				if (!Hasil) return console.log(Hasil)
-				const Respon: { sender: string, id: number, filesha: string, mess: proto.WebMessageInfo} = Hasil.find((value) => value.filesha === FileSha)
-				await this.Ra.sendTextWithMentions(from, StickerDuplicate(Respon.sender, Respon.id), [Respon.sender], mess)
-				await this.Ra.sendTextWithMentions(from, StickerFound(sender), [sender], Respon.mess)
+				const Respon: { sender: string, id: number, filesha: string, mess: proto.WebMessageInfo} | undefined = Hasil.find((value) => value.filesha === FileSha)
+				await this.Ra.sendTextWithMentions(from, StickerDuplicate(Respon?.sender || "", Respon?.id || 0), [Respon?.sender || ""], mess)
+				await this.Ra.sendTextWithMentions(from, StickerFound(sender || ""), [sender || ""], Respon?.mess)
 			}
-			if (CheckSticker(from, FileSha, Stick)) return
-			const Wm: string = args[0] === undefined ? undefined : args.join(" ")
+			if (CheckSticker(from, FileSha || "", Stick)) return
+			const Wm: string | null | undefined = args[0] === undefined ? undefined : args.join(" ")
 			await this.Ra.reply(from, IndTunggu(), mess)
 			const input: string = await res.downloadAndSaveMediaMessage(media,  "./lib/storage/temp/" + RandomName(22))
 			await CreateSticker(input, Wm).then(async (result: string | Error) => {
 				try {
 					if (typeof result !== "string") return
 					let ress: proto.WebMessageInfo  = await res.sendMessage(from,  fs.readFileSync(result), MessageType.sticker, { quoted: mess})
-					await AddSticker(ress, from, FileSha, sender, Stick)
+					await AddSticker(ress, from, FileSha || "", sender || "", Stick)
 					await Tunggu(2000)
 					if (fs.existsSync(result)) fs.unlinkSync(result)
 					if (fs.existsSync(input)) fs.unlinkSync(input)
@@ -50,7 +50,7 @@ export class Converter {
 					try {
 						if (typeof Res !== "string") return
 						let ress: proto.WebMessageInfo  = await res.sendMessage(from,  fs.readFileSync(Res), MessageType.sticker, { quoted: mess})
-						await AddSticker(ress, from, FileSha, sender, Stick)
+						await AddSticker(ress, from, FileSha || "", sender || "", Stick)
 						await Tunggu(2000)
 						if (fs.existsSync(Res)) fs.unlinkSync(Res)
 						if (fs.existsSync(input)) fs.unlinkSync(input)
@@ -63,7 +63,7 @@ export class Converter {
 						try {
 							if (typeof Result !== "string") return
 							let ress: proto.WebMessageInfo  = await res.sendMessage(from,  fs.readFileSync(Result), MessageType.sticker, { quoted: mess})
-							await AddSticker(ress, from, FileSha, sender, Stick)
+							await AddSticker(ress, from, FileSha || "", sender || "", Stick)
 							await Tunggu(2000)
 							if (fs.existsSync(Result)) fs.unlinkSync(Result)
 							if (fs.existsSync(input)) fs.unlinkSync(input)
@@ -76,7 +76,7 @@ export class Converter {
 							try {
 								if (typeof Res !== "string") return
 								let ress: proto.WebMessageInfo = await res.sendMessage(from,  fs.readFileSync(Res), MessageType.sticker, { quoted: mess})
-								await AddSticker(ress, from, FileSha, sender, Stick)
+								await AddSticker(ress, from, FileSha || "", sender || "", Stick)
 								await Tunggu(2000)
 								if (fs.existsSync(Res)) fs.unlinkSync(Res)
 								if (fs.existsSync(input)) fs.unlinkSync(input)
@@ -99,12 +99,13 @@ export class Converter {
 			const { isQuotedSticker, media, from, mess, sender } = data
 			if (!media && !isQuotedSticker)  return this.Ra.reply(from, IndBukanSticker(), mess)
 			await this.Ra.reply(from, IndTunggu(), mess)
-			const input: string = await res.downloadAndSaveMediaMessage(media, "./lib/storage/temp/" + RandomName(23) + sender.replace("@s.whatsapp.net", ""))
-			await ToVideo(input).then(async (value: { status: number, data: string }) => {
-				await this.Ra.sendVideo(from, await Buffer(value.data), "", mess)
+			if (!media) return
+			const input: string = await res.downloadAndSaveMediaMessage(media, "./lib/storage/temp/" + RandomName(23) + sender?.replace("@s.whatsapp.net", ""))
+			await ToVideo(input).then(async (value: { status: number, data: string } | any) => {
+				await this.Ra.sendVideo(from, await Buffer(value?.data), "", mess)
 				if (fs.existsSync(input)) fs.unlinkSync(input)
 			}).catch(async (err: Error) => {
-				let respon: string
+				let respon: string | any
 				try {
 					respon = await toVideoV2(input)
 				} catch (err) {
@@ -123,6 +124,7 @@ export class Converter {
 		globalThis.CMD.on("converter|tomp3", "tomp3", async (res: WAConnection, data:  Commands) => {
 			const { media, isQuotedVideo, from, mess, isVideo, sender, sendOwner } = data
 			if (!media && !isQuotedVideo || !media && isVideo) return this.Ra.reply(from, IndBukanVid(), mess)
+			if (!media) return
 			await this.Ra.reply(from, IndTunggu(),mess)
 			const input: string = await res.downloadAndSaveMediaMessage(media, "./lib/storage/temp/" +  RandomName(24))
 			await Tomp3(input).then(async (result: string | Error) => {
@@ -146,6 +148,7 @@ export class Converter {
 	protected async toImg () {
 		globalThis.CMD.on("converter|toimg", "toimg", async (res: WAConnection, data: Commands) => {
 			const { from, mess} = data
+			if (!data.media) return
 			const result =  await  Toimg(await res.downloadAndSaveMediaMessage(data.media,"./lib/storage/temp/" + RandomName(27)))
 			await this.Ra.sendImage(from, fs.readFileSync(result || ""), "", mess)
 		}, { noPrefix: false})
@@ -154,6 +157,7 @@ export class Converter {
 		globalThis.CMD.on("converter|tocute", "tocute", async (res: WAConnection, data:  Commands) => {
 			const { media, isQuotedAudio, from, mess, isAudio, sender } = data
 			if (!media && !isQuotedAudio || !media && isAudio) return this.Ra.reply(from, IndBukanAud(), mess)
+			if (!media) return
 			await this.Ra.reply(from, IndTunggu(), mess)
 			const input: string = await res.downloadAndSaveMediaMessage(media, "./lib/storage/temp/" +  RandomName(28))
 			await Tocute(input).then(async (result: string | Error) => {
